"공통응답 및 예외처리 적용"

+++ src/main/java/com/nbcamp/mypocketbookapi/service/CommentService.java
@@ -5,6 +5,8 @@ import com.nbcamp.mypocketbookapi.dto.comment.response.CommentResponse;
 import com.nbcamp.mypocketbookapi.entity.Comment;
 import com.nbcamp.mypocketbookapi.entity.Member;
 import com.nbcamp.mypocketbookapi.entity.Review;
+import com.nbcamp.mypocketbookapi.exception.ErrorCode;
+import com.nbcamp.mypocketbookapi.exception.comment.CommentException;
 import com.nbcamp.mypocketbookapi.repository.CommentJpaRepository;
 import com.nbcamp.mypocketbookapi.repository.MemberJpaRepository;
 import com.nbcamp.mypocketbookapi.repository.ReviewJpaRepository;
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 
@@ -28,11 +31,11 @@ public class CommentService {
 
 
     @Transactional
-    public Comment createComment(CommentRequest request) {
-        Member member = memberJpaRepository.findById(request.getMemberId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
-        Review review = reviewJpaRepository.findById(request.getReviewId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 리뷰입니다."));
+    public Comment createComment(Long reviewId, Long memberId, CommentRequest request) {
+        Member member = memberJpaRepository.findById(memberId)
+                .orElseThrow(() -> new CommentException(ErrorCode.MEMBER_NOT_FOUND));
+        Review review = reviewJpaRepository.findById(reviewId)
+                .orElseThrow(() -> new CommentException(ErrorCode.REVIEW_NOT_FOUND));
 
         Comment comment = Comment.builder()
                 .member(member)
@@ -45,7 +48,7 @@ public class CommentService {
     @Transactional
     public List<CommentResponse> getCommentsByReviewId(Long reviewId) {
         Review review = reviewJpaRepository.findById(reviewId)
-                .orElseThrow(() -> new EntityNotFoundException("리뷰가 존재하지 않습니다."));
+                .orElseThrow(() -> new CommentException(ErrorCode.REVIEW_NOT_FOUND));
 
         List<Comment> comments = commentJpaRepository.findByReview(review);
 
@@ -55,16 +58,26 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(Long id, String newText) {
-        Comment comment = commentJpaRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
+    public void updateComment(Long commentId, Long memberId, String newText) {
+        Comment comment = commentJpaRepository.findById(commentId)
+                .orElseThrow(() -> new CommentException(ErrorCode.CONTENT_NOT_FOUND));
+        Member member = memberJpaRepository.findById(memberId)
+                .orElseThrow(() -> new CommentException(ErrorCode.MEMBER_NOT_FOUND));
+        if(!Objects.equals(comment.getMember().getId(), member.getId())) {
+            throw new CommentException(ErrorCode.UNAUTHORIZED_REVIEW_MODIFICATION);
+        }
         comment.updateText(newText);
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long memberId) {
         Comment comment = commentJpaRepository.findById(commentId)
-                .orElseThrow(() -> new EntityNotFoundException("댓글이 존재하지 않습니다."));
+                .orElseThrow(() -> new CommentException(ErrorCode.CONTENT_NOT_FOUND));
+        Member member = memberJpaRepository.findById(memberId)
+                .orElseThrow(() -> new CommentException(ErrorCode.MEMBER_NOT_FOUND));
+        if(!Objects.equals(comment.getMember().getId(), member.getId())) {
+            throw new CommentException(ErrorCode.UNAUTHORIZED_COMMENT_DELETION);
+        }
         commentJpaRepository.delete(comment);
     }
 }
